Ajustes
1. Adición de emojis en mensajes mostrados en consola.

2. Cambio de valores en el WaitForSeconds para evitar la sobrecarga.

3. Ajuste de botones y panel de la GUI.

4. Cambio en el tipo de dato del atributo tiempo_promedio_despacho en la clase ExportJson

+++ Ejemplo_GITG52/Taller1_G52/Assets/Scripts/Export JSON.cs
@@ -8,7 +8,7 @@ public class ExportJSON : MonoBehaviour
     public int total_generado;        
     public int total_despachados;   
     public int total_pila;           
-    public int tiempo_promedio_despacho;
+    public float tiempo_promedio_despacho;
     public Dictionary<string, int> despachadosPorTipo;
     public string tipoMasDespachado;
 }
@@ -36,7 +36,7 @@ public class Exportador
         public int total_generado;         
         public int total_despachados;      
         public int total_pila;           
-        public int tiempo_promedio_despacho;
+        public float tiempo_promedio_despacho;
         public List<TipoCantidad> despachadosPorTipo;
         public string tipoMasDespachado;
